// src/app/shared/interfaces/discount/discount.interface.ts
export interface IDiscount {
    id: number;
    name: string;
    title: string;
    description: string;
    imagePath: string;
}

export interface IDiscountRequest {
    name: string;
    title: string;
    description: string;
    imagePath: string;
}

export interface IDiscountResponse extends IDiscountRequest {
    id: number | string;
}